# optimized polynomial solvers