--- a/kornia/geometry/solvers/polynomial_solver.py
+++ b/kornia/geometry/solvers/polynomial_solver.py
@@ -243,7 +243,7 @@ T_deg1[14, 8] = 1  # 1 * z
 T_deg1[15, 9] = 1  # 1 * 1
 
 
-def multiply_deg_one_poly(a: torch.Tensor, b: torch.Tensor, T_matrix: torch.Tensor = T_deg1) -> torch.Tensor:
+def multiply_deg_one_poly(a: torch.Tensor, b: torch.Tensor) -> torch.Tensor:
     r"""Multiply two polynomials of the first order [@nister2004efficient].
 
     Args:
@@ -254,9 +254,10 @@ def multiply_deg_one_poly(a: torch.Tensor, b: torch.Tensor, T_matrix: torch.Tens
         degree 2 poly with the order :math:`(x^2, x*y, x*z, x, y^2, y*z, y, z^2, z, 1)`.
 
     """
-    if T_matrix.device != a.device or T_matrix.dtype != a.dtype:
-        T_matrix = T_matrix.to(device=a.device, dtype=a.dtype)
-    return (a.unsqueeze(2) * b.unsqueeze(1)).flatten(start_dim=-2) @ T_matrix
+    global T_deg1 # noqa: PLW0603
+    if T_deg1.device != a.device or T_deg1.dtype != a.dtype:
+        T_deg1 = T_deg1.to(device=a.device, dtype=a.dtype)
+    return (a.unsqueeze(2) * b.unsqueeze(1)).flatten(start_dim=-2) @ T_deg1
 
 
 # Reference
@@ -306,7 +307,7 @@ T_deg2[38, 18] = 1  # (9*4+2)
 T_deg2[39, 19] = 1  # (9*4+3)
 
 
-def multiply_deg_two_one_poly(a: torch.Tensor, b: torch.Tensor, T_matrix: torch.Tensor = T_deg2) -> torch.Tensor:
+def multiply_deg_two_one_poly(a: torch.Tensor, b: torch.Tensor) -> torch.Tensor:
     r"""Multiply two polynomials a and b of degrees two and one [@nister2004efficient].
 
     Args:
@@ -319,11 +320,12 @@ def multiply_deg_two_one_poly(a: torch.Tensor, b: torch.Tensor, T_matrix: torch.
         x*y*z, x*y, x*z^2, x*z, x, y*z^2, y*z, y, z^3, z^2, z, 1)`.
 
     """
-    if T_matrix.device != a.device or T_matrix.dtype != a.dtype:
-        T_matrix = T_matrix.to(device=a.device, dtype=a.dtype)
+    global T_deg2 # noqa: PLW0603
+    if T_deg2.device != a.device or T_deg2.dtype != a.dtype:
+        T_deg2 = T_deg2.to(device=a.device, dtype=a.dtype)
     product_basis = a.unsqueeze(2) * b.unsqueeze(1)
     product_vector = product_basis.flatten(start_dim=-2)
-    return product_vector @ T_matrix
+    return product_vector @ T_deg2
 
 
 # Compute degree 10 poly representing determinant (equation 14 in the paper)
@@ -1793,9 +1795,6 @@ coefficient_map = torch.tensor(
 
 def determinant_to_polynomial(
     A: Tensor,
-    multiplication_indices: Tensor = multiplication_indices,
-    signs: Tensor = signs,
-    coefficient_map: Tensor = coefficient_map,
 ) -> Tensor:
     r"""Represent the determinant by the 10th polynomial, used for 5PC solver [@nister2004efficient].
 
@@ -1807,6 +1806,7 @@ def determinant_to_polynomial(
 
     """
     B, device, dtype = A.shape[0], A.device, A.dtype
+    global multiplication_indices, signs, coefficient_map # noqa: PLW0603
 
     multiplication_indices = multiplication_indices.to(device)
     signs = signs.to(device, dtype)