@sidd-27 2025-10-05
# noqa: PLW0603 has been used to supress the ruff errors, as the matrix isnt designed to change under any operation, if theres a better practice on how to handle global variables pls do let me know